chore: abstract obs date formatter & move formatters to single file
